requirements.txt
httpx==0.27.2
openai==1.55.3
vosk==0.3.45
PyAudio==0.2.14
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5-sip==12.16.1
open-interpreter==0.4.3
pyttsx3==2.98
